Refactor slices in `forward()` methods.
I was finding it hard to read the slice notation in the
`FactorizedSpectralConvNd.forward` implementations (i.e. where
`out_fft[*] = self._contract(x[*], ...)`) for what's supposed to be the same
slice `[*]`. And autopep8 was mangling that part.

Refactor each call to `self._contract` to use the same slice to index each
tensor (i.e. to index `out_fft` and `x`).